requirements.txt
-i https://pypi.org/simple
alembic==1.13.1; python_version >= '3.8'
apiclient==1.0.4
argcomplete==3.2.1; python_version >= '3.8'
blinker==1.7.0; python_version >= '3.8'
cachetools==5.3.2; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
cfgv==3.4.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cloudinary==1.38.0
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
cryptography==41.0.7; python_version >= '3.7'
distlib==0.3.8
filelock==3.13.1; python_version >= '3.8'
flask==3.0.0; python_version >= '3.8'
flask-admin==1.6.1; python_version >= '3.6'
flask-cors==4.0.0
flask-migrate==4.0.5; python_version >= '3.6'
flask-sqlalchemy==3.0.5; python_version >= '3.7'
flask-swagger==0.2.14
google-api-core==2.15.0; python_version >= '3.7'
google-api-python-client==2.113.0; python_version >= '3.7'
google-auth==2.26.2; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0; python_version >= '3.6'
googleapis-common-protos==1.62.0; python_version >= '3.7'
greenlet==3.0.3; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==21.2.0; python_version >= '3.5'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
identify==2.5.33; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.3; python_version >= '3.7'
jsonify==0.5
mako==1.3.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
nodeenv==1.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
oauthlib==3.2.2; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pipx==1.4.2; python_version >= '3.8'
platformdirs==4.1.0; python_version >= '3.8'
pre-commit==3.6.0; python_version >= '3.9'
protobuf==4.25.2; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyasn1==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pyparsing==3.1.1; python_version >= '3.1'
python-dotenv==1.0.0; python_version >= '3.8'
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9; python_version >= '3.6' and python_version < '4'
setuptools==69.0.3; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.46; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
todoist-api-python==2.1.3; python_version >= '3.9' and python_version < '4.0'
tomli==2.0.1; python_version < '3.11'
typing-extensions==4.9.0; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.1.0; python_version >= '3.8'
userpath==1.9.1; python_version >= '3.7'
virtualenv==20.25.0; python_version >= '3.7'
werkzeug==3.0.1; python_version >= '3.8'
wtforms==3.1.2; python_version >= '3.8'
